QPID-3858: Updated branch - merged from trunk r.1333987
git-svn-id: https://svn.apache.org/repos/asf/qpid/branches/asyncstore@1334037 13f79535-47bb-0310-9956-ffa450edef68

+++ java/bdbstore/src/main/java/org/apache/qpid/server/store/berkeleydb/AMQShortStringEncoding.java
@@ -33,7 +33,7 @@ public class AMQShortStringEncoding
 
     public static AMQShortString readShortString(TupleInput tupleInput)
     {
-        int length = (int) tupleInput.readShort();
+        int length = tupleInput.readShort();
         if (length < 0)
         {
             return null;
